These are doubles, not ints!

+++ libmysolvers.c
@@ -70,7 +70,7 @@ double linear_solver(int val, double a1, double b1, double c1, double a2, double
     }
 }
 
-double triangle_centroid(int val, int x1, int y1, int x2, int y2, int x3, int y3){
+double triangle_centroid(int val, double x1, double y1, double x2, double y2, double x3, double y3){
     double centroid_x = (x1 + x2 + x3) / 3.0;
     double centroid_y = (y1 + y2 + y3) / 3.0;
